# perf: faster local attr lookup