#### Before

```py
In [2]: %timeit -n 1_000_000 frappe.db.get_value
512 ns ± 4.74 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)

In [3]: %timeit -n 1_000_000 frappe.local.db.get_value
348 ns ± 3.28 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
```

#### After

```py
In [1]: %timeit -n 1_000_000 frappe.db.get_value
263 ns ± 2.9 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)

In [2]: %timeit -n 1_000_000 frappe.local.db.get_value
144 ns ± 1.5 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
```